backend/requirements.txt
fastapi==0.103.2
uvicorn[standard]==0.23.2
sqlalchemy==2.0.22
pydantic[email]==2.4.2
pydantic-settings==2.0.3
python-jose[cryptography]==3.3.0
pycryptodome==3.19.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
psycopg2-binary==2.9.9
pytest==7.4.2
google-generativeai==0.7.2
